@@ -4,14 +4,14 @@ version: 1.0.0
 homepage: https://github.com/git-touch/github-trending
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  http: ^0.12.0
-  json_annotation: ^3.0.1
+  http: ^0.13.0
+  json_annotation: ^4.0.0
 
 dev_dependencies:
-  pedantic: ^1.0.0
-  test: ^1.0.0
-  build_runner: ^1.7.4
-  json_serializable: ^3.2.5
+  pedantic: ^1.11.0
+  test: ^1.16.5
+  build_runner: ^1.11.5
+  json_serializable: ^4.0.2
