@@ -4,17 +4,17 @@ part 'model.g.dart';
 
 @JsonSerializable()
 class GithubTrendingRepository {
-  String author;
-  String name;
-  String avatar;
-  String url;
-  String description;
-  String language;
-  String languageColor;
-  int stars;
-  int forks;
-  int currentPeriodStars;
-  List<GithubTrendingRepositoryBuiltBy> builtBy;
+  String? author;
+  String? name;
+  String? avatar;
+  String? url;
+  String? description;
+  String? language;
+  String? languageColor;
+  int? stars;
+  int? forks;
+  int? currentPeriodStars;
+  List<GithubTrendingRepositoryBuiltBy>? builtBy;
   GithubTrendingRepository();
   factory GithubTrendingRepository.fromJson(Map<String, dynamic> json) =>
       _$GithubTrendingRepositoryFromJson(json);
@@ -22,9 +22,9 @@ class GithubTrendingRepository {
 
 @JsonSerializable()
 class GithubTrendingRepositoryBuiltBy {
-  String href;
-  String avatar;
-  String username;
+  String? href;
+  String? avatar;
+  String? username;
   GithubTrendingRepositoryBuiltBy();
   factory GithubTrendingRepositoryBuiltBy.fromJson(Map<String, dynamic> json) =>
       _$GithubTrendingRepositoryBuiltByFromJson(json);
@@ -32,12 +32,12 @@ class GithubTrendingRepositoryBuiltBy {
 
 @JsonSerializable()
 class GithubTrendingDeveloper {
-  String username;
-  String name;
-  String type;
-  String url;
-  String avatar;
-  GithubTrendingDeveloperRepository repo;
+  String? username;
+  String? name;
+  String? type;
+  String? url;
+  String? avatar;
+  GithubTrendingDeveloperRepository? repo;
   GithubTrendingDeveloper();
   factory GithubTrendingDeveloper.fromJson(Map<String, dynamic> json) =>
       _$GithubTrendingDeveloperFromJson(json);
@@ -45,9 +45,9 @@ class GithubTrendingDeveloper {
 
 @JsonSerializable()
 class GithubTrendingDeveloperRepository {
-  String name;
-  String description;
-  String url;
+  String? name;
+  String? description;
+  String? url;
   GithubTrendingDeveloperRepository();
   factory GithubTrendingDeveloperRepository.fromJson(
           Map<String, dynamic> json) =>
